# swagger.yml
openapi: 3.0.0
info:
  title: Character API
  version: 1.0.0
  description: API for managing characters
paths:
  /character:
    post:
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                nickname:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                active:
                  type: boolean
              required:
                - name
                - surname
                - nickname
                - email
                - phone
                - address
                - active
      responses:
        '400':
          description: Bad arguments
        '201':
          description: Character created
        '500':
          description: Unknown error
  /character/{characterId}:
    get:
      summary: Retrieve a character
      parameters:
        - in: path
          name: characterId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Character retrieved
        '400':
          description: Bad arguments
        '404':
          description: Character not found
        '500':
          description: Unknown error
    patch:
      summary: Update a character
      parameters:
        - in: path
          name: characterId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                nickname:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                active:
                  type: boolean
      responses:
        '204':
          description: Character updated
        '400':
          description: Bad arguments
        '404':
          description: Character not found
        '500':
          description: Unknown error
    delete:
      summary: Delete a character
      parameters:
        - in: path
          name: characterId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Character deleted
        '400':
          description: Bad arguments
        '404':
          description: Character not found
        '500':
          description: Unknown error